+++ README.md
@@ -1,7 +1,8 @@
 ![Main icon](http://goo.gl/5HcH7T)
+On travis the build is: [![Build Status](https://travis-ci.org/boomboompower/SimpleChatAlert.svg?branch=master)](https://travis-ci.org/boomboompower/SimpleChatAlert)
 
-Whats it for | New features | News about it
------------- | ------------- | -------------
-What we think is one of the best uses for this plugin is, Alerting players if you are holding an event, e.g: _/alert We are holding a drop party at spawn!_ would become: ALERT: We are holding a drop party at spawn! This is a clear message in the chat which people will be able to see easily. | It will probably be updated every time minecraft updates for a couple of versions, then I will re-introduce dependencies so I don't have to maintain it through updating NMS anymore. | It will probably be updated every time minecraft updates for a couple of versions, then I will re-introduce dependencies so I don't have to maintain it through updating NMS anymore.
+Whats it for | New features 
+------------ | -------------
+What we think is one of the best uses for this plugin is, Alerting players if you are holding an event, e.g: _/alert We are holding a drop party at spawn!_ would become: ALERT: We are holding a drop party at spawn! This is a clear message in the chat which people will be able to see easily. | All features including Chat, ActionBars, Titles, PlayerList header/footer alerts are now 110% built in! No un-needed dependancies!
 
 Downloads located on the [_Archives_](http://boomboompower.weebly.com/archives.html) page
