--- SimpleChatAlert v9/AlertSettings/ActionBar.yml
@@ -1,6 +1,0 @@
-# No API is needed for this feature its now 110% built in!
-ActionBar: false
-
-Prefix: '&4&lALERT &c&l'
-
-Suffix: ''
