--- SimpleChatAlert v9/AlertSettings/Title.yml
@@ -1,10 +1,0 @@
-# Allows you to make the alert be in the form of a title   #
-# Change TitlePrefix to change the top bar and edit        #
-# TitleColor to edit the subtitle no API needed for this!  #        
-Title: false
-
-# Change the title's color/format, must have '' in it.     #
-TitlePrefix: '&4&lALERT:'
-
-# Change the subtitles color/format, must have '' in it    #
-TitleColor: '&c&l'
