--- SimpleChatAlert v9/Languages/Lang.yml
@@ -1,3 +1,0 @@
-NoPermission: '&cIm sorry, but you do not have permission to perform this command. Please contact the server administrators if you believe that this is in error.'
-
-NoMessage: ''
