--- SimpleChatAlert v9/README.yml
@@ -1,23 +1,0 @@
-# Hey! Thanks a lot for downloading SimpleChatAlert!
-# We hope you have a great experience using it.
-# This is just a automatically generated README document. 
-# Removing it will regenerate it when the plugin reloads
-# This plugin was created by boomboompower
-#
-# -- Terms Of Use --
-# Do not upload this plugin to a website without permission
-# If you are going to make a custom version, message me https://bukkit.org/conversations/add?to=boomboompower
-#
-# Pretty simple eh ;)
-#
-# -- Links --
-# The source code is availible at: https://github.com/boomboompower/SimpleChatAlert/
-# My official website: http://boomboompower.weebly.com/
-# Found any bugs? Report them here: http://dev.bukkit.org/bukkit-plugins/simple-chat-alert/
-# Our Spigot Page: https://www.spigotmc.org/resources/simplechatalert.13563/
-#
-# -- More Info --
-# If you find anyone redistributing the plugin without permission, report them with evidence.
-# Thank you for using BlockHandler and please remember to give feedback! :D
-#
-#
