--- SimpleChatAlert v9/config.yml
@@ -1,55 +1,0 @@
-#|---------------------------------------------|#
-#|  #####  ####  #     # ###### #######  ##### |#
-#| #      #    # ##    # #         #    #      |#
-#| #      #    # # #   # #         #    #      |#
-#| #      #    # #  #  # ####      #    #   ## |#
-#| #      #    # #   # # #         #    #    # |#
-#| #      #    # #    ## #         #    #    # |#
-#|  #####  ####  #     # #      #######  ##### |#
-#|---------------------------------------------|#
-
-Language: en_US
-
-# Change what the message looks like! Don't forget the ''  #
-# You can use {PLAYER} which puts the senders username!    #
-# You can also use color codes by using & -Enjoy SCA       #
-Message: '&4&lALERT: &c&l'
-
-# Change what the mesage is if the sender does not have    #
-# the correct permission. You can use color codes.
-NoPermission: '&4You do not have access to this command!'
-
-# Change what the message is if the sender does not imput  #
-# A message with the /alert command. Uses color codes.     #
-NoMessage: '&cPlease add a message! Example: &4&o/alert Welcome to the server!'
-
-# Do you want the command to put the message into the chat #
-ChatAlert: true
-
-ActionBar: #Now located in ActionBar.yml
-BossBar: #Now located in BossBar.yml 
-Title: #Now located in Title.yml
-
-# Do you want the alert to show up on the PlayerList?      #
-# This is also known as the "Header and Footer", enjoy!    #
-PlayerList: false
-
-# What do you want to show up on the Header/PlayerListTop  #
-PlayerListHeader: '&4&lALERT'
-
-# How do you want the format on Footer/Bottom to be?       #
-PlayerListFooter: '&c&l'
-
-# Do you want the sender to `Glow`? - EXPERIMENTAL FEATURE #
-Glow: false
-
-# Done in seconds, how long should the player glow?        #
-GlowTime: 10
-
-# Checks if there are any updates and messages the console #
-# If any are found, it will put a message in the console   #
-# Throws an error if you are not connected to the internet #
-UpdateChecker: false
-
-# Send anonymous statistics to http://mcstats.org its safe #
-MetricsLite: true
