--- SimpleChatAlert v9/plugin.yml
@@ -1,16 +1,0 @@
-name: SimpleChatAlert
-version: 9.1
-main: me.boomboompower.SimpleChatAlert
-author: boomboompower
-description: A very simple lightweight alert plugin!
-prefix: SCA
-website: dev.bukkit.org/bukkit-plugins/simple-chat-alert
-commands:
-   alert:
-      description: Sends an alert to the server
-      usage: /<command> [Message]
-      default: op
-   sca:
-      description: Debugging and options for SCA
-      usage: /<command> help
-      default: op
