+++ src/main/resources/plugin.yml
@@ -1,0 +1,12 @@
+name: SimpleChatAlert
+version: ${project.version}
+main: me.boomboompower.sca.SimpleChatAlert
+author: boomboompower
+description: A very simple lightweight alert plugin!
+prefix: SCA
+website: dev.bukkit.org/bukkit-plugins/simple-chat-alert
+commands:
+   alert:
+      description: Sends an alert to the server
+      usage: /<command> <message>
+      default: op
